+++ griptape/engines/__init__.py
@@ -12,6 +12,7 @@ from .image.inpainting_image_generation_engine import InpaintingImageGenerationE
 from .image.outpainting_image_generation_engine import OutpaintingImageGenerationEngine
 from .image_query.image_query_engine import ImageQueryEngine
 from .audio.text_to_speech_engine import TextToSpeechEngine
+from .audio.audio_transcription_engine import AudioTranscriptionEngine
 
 __all__ = [
     "BaseQueryEngine",
@@ -28,4 +29,5 @@ __all__ = [
     "OutpaintingImageGenerationEngine",
     "ImageQueryEngine",
     "TextToSpeechEngine",
+    "AudioTranscriptionEngine",
 ]